app/requirements.txt
Pillow==8.2.0
folium==0.12.1
Flask
psycopg2-binary
pandas